// src/models/UserProfile/UserProfileFetched.ts
export interface UserProfileFetched {
    id:            number;
    username:      string;
    password:      string;
    email:         string;
    full_name:     string;
    phone:         string;
    address:       string;
    avatar:        string;
    is_active:     boolean;
    role:          string;
    create_date:   null;
    update_date:   null;
}